feat: Rename dir ( the rest components are widgets )
