+++ src/widgets/column/column.scss
