+++ src/widgets/editor/editor.jsx
