+++ src/widgets/editor/editor.scss
