+++ src/widgets/layout/layout.scss
