+++ src/widgets/menu/menu.scss
