+++ src/widgets/sidebar/dialogs.jsx
