+++ src/widgets/sidebar/dialogs.scss
