+++ src/widgets/sidebar/sidebar.scss
